requirements.txt
pandas==2.2.0
openpyxl==3.1.2
python-docx==1.1.0
python-pptx==0.6.23
PyPDF2==3.0.1
markdown==3.5.2
pyyaml==6.0.1
jinja2==3.1.3
